Refactoring and deleting Main/Greeting
