+++ app/src/androidTest/java/com/github/polypoly/app/JoinGroupActivityTest.kt
@@ -3,6 +3,7 @@ package com.github.polypoly.app
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.polypoly.app.menu.JoinGroupActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
