+++ app/src/main/AndroidManifest.xml
@@ -16,40 +16,32 @@
         android:theme="@style/Theme.Polypoly"
         tools:targetApi="31">
         <activity
-            android:name=".JoinGroupActivity"
+            android:name=".MapActivity"
+            android:exported="true" />
+
+        <!-- ========================================= MENU ACTIVITIES -->
+        <activity
+            android:name=".menu.JoinGroupActivity"
             android:exported="false"
             android:label="@string/title_activity_join_group"
             android:theme="@style/Theme.Polypoly" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".menu.SettingsActivity"
             android:exported="false"
             android:label="@string/title_activity_settings"
             android:theme="@style/Theme.Polypoly" />
         <activity
-            android:name=".MapActivity"
-            android:exported="true" />
-        <activity
-            android:name=".GreetingActivity"
-            android:exported="false"
-            android:label="@string/title_activity_greeting"
-            android:theme="@style/Theme.Polypoly" />
-        <activity
-            android:name=".ProfileModifyingActivity"
+            android:name=".menu.ProfileModifyingActivity"
             android:exported="true"
             android:label="@string/title_activity_greeting"
             android:theme="@style/Theme.Polypoly" />
         <activity
-            android:name=".ProfileActivity"
+            android:name=".menu.ProfileActivity"
             android:exported="true"
             android:label="@string/title_activity_greeting"
             android:theme="@style/Theme.Polypoly" />
         <activity
-            android:name=".MainActivity"
-            android:exported="false"
-            android:label="@string/title_activity_greeting"
-            android:theme="@style/Theme.Polypoly" />
-        <activity
-            android:name=".WelcomeActivity"
+            android:name=".menu.WelcomeActivity"
             android:exported="true"
             android:label="@string/title_activity_welcome"
             android:theme="@style/Theme.Polypoly">
@@ -59,6 +51,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <!-- ========================================================= -->
+
     </application>
-    <uses-permission android:name="android.permission.READ_TOASTS" />
+    <uses-permission android:name="android.permission.READ_TOASTS"
+        tools:ignore="ManifestOrder" />
 </manifest>
