--- app/src/main/java/com/github/polypoly/app/GreetingActivity.kt
@@ -1,96 +1,0 @@
-package com.github.polypoly.app
-
-import android.content.res.Configuration
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.tooling.preview.Preview
-
-import com.github.polypoly.app.ui.theme.PolypolyTheme
-
-class GreetingActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            PolypolyTheme {
-                // This is the surface where all the view lies
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colors.background
-                ) {
-                    Column(
-                        modifier = Modifier.fillMaxSize(),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        // The name is sent in the intent
-                        val name = intent.getStringExtra("name")
-                        if (name != null) {
-                            GreetingText(name)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * A simple view that displays a greeting message to a given name
- */
-@Composable
-fun GreetingText(name: String) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            modifier = Modifier.testTag("greetingText"),
-            text = "Good morning $name",
-            style = MaterialTheme.typography.h6
-        )
-        Text(
-            text = "You look beautiful today <3",
-            style = MaterialTheme.typography.body2
-        )
-    }
-}
-
-
-
-
-
-
-
-// =================================== PREVIEW ==============
-@Preview(name = "Light Mode")
-@Preview(
-    name = "Dark Mode",
-    uiMode = Configuration.UI_MODE_NIGHT_YES
-)
-@Composable
-fun GreetingPreview() {
-    PolypolyTheme {
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colors.background
-        ) {
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                GreetingText(name = "bigflo")
-            }
-        }
-    }
-}
